requirements.txt
DateTime==4.3
entrypoints==0.3
Flask==1.1.2
idna==2.9
importlib-metadata==1.6.0
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.25
pygame==2.0.1
requests==2.23.0
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.9
